MAINT: Reorganize readers into IO. (#26)
