+++ pydisdrometer/__init__.py
@@ -1,14 +1,16 @@
-from .ParsivelReader import read_parsivel
-from .ParsivelNasaGVReader import read_parsivel_nasa_gv
-from .JWDReader import read_jwd
+from .io.ParsivelReader import read_parsivel
+from .io.ParsivelNasaGVReader import read_parsivel_nasa_gv
+from .io.JWDReader import read_jwd
+from .io.Image2DReader import read_ucsc_netcdf, read_noaa_aoml_netcdf
+
 from .aux_readers.GPMApuWallopsRawReader import read_gpm_nasa_apu_raw_wallops
 from .aux_readers.NASA_2DVD_reader import read_2dvd_sav_nasa_gv
 from .aux_readers.NASA_2DVD_reader import read_2dvd_dsd_nasa_gv
 from .aux_readers.ARM_APU_reader import read_parsivel_arm_netcdf
-from .plot.plot import plot_dsd
 from .aux_readers.read_2ds import read_2ds
 from .aux_readers.read_hvps import read_hvps
-from .io.Image2DReader import read_ucsc_netcdf, read_noaa_aoml_netcdf
+
+from .plot.plot import plot_dsd
 
 from . import partition
 from . import utility
